+++ src/layers/l4-controller/commands/file/index.ts
@@ -1,0 +1,2 @@
+export { UpdateAttachmentDownloadCommand } from './UpdateAttachmentDownloadCommand';
+export { ClearSyncedFilesCommand } from './ClearSyncedFilesCommand';