+++ src/layers/l4-controller/commands/index.ts
@@ -15,3 +15,4 @@ export * from './export';
 export * from './resource';
 export * from './syncUpdate';
 export * from './htmlDependency';
+export * from './file';